+++ src/pipecatcloud/__init__.py
@@ -7,6 +7,5 @@ PIPECAT_CLI_NAME = "pipecat"
 PIPECAT_CREDENTIALS_PATH = "~/.pipecatcloud.toml"
 PIPECAT_DEPLOY_CONFIG_PATH = "pcc-deploy.toml"
 
-
 logger.remove()
 logger.add(sys.stderr, level=str(os.getenv("PCC_LOG_LEVEL", "INFO")))