+++ src/pipecatcloud/_utils/console_utils.py
@@ -1,25 +1,71 @@
-
 from typing import Optional
 
 from rich.console import Console
 from rich.panel import Panel
 
-from pipecatcloud.cli import PANEL_TITLE_ERROR
+from pipecatcloud.cli import PANEL_TITLE_ERROR, PANEL_TITLE_SUCCESS
 from pipecatcloud.errors import ERROR_CODES
 
-console = Console()
+
+class PipecatConsole(Console):
+    def success(
+            self,
+            message,
+            title_extra: Optional[str] = None,
+            subtitle: Optional[str] = None):
+        title = f"{PANEL_TITLE_SUCCESS}{f' - {title_extra}' if title_extra is not None else ''}"
+
+        self.print(
+            Panel(
+                message,
+                title=f"[bold green]{title}[/bold green]",
+                subtitle=subtitle,
+                title_align="left",
+                subtitle_align="left",
+                border_style="green"))
+
+    def error(
+            self,
+            message,
+            title_extra: Optional[str] = None,
+            subtitle: Optional[str] = None):
+
+        title = f"{PANEL_TITLE_ERROR}{f' - {title_extra}' if title_extra is not None else ''}"
+
+        self.print(
+            Panel(
+                message,
+                title=f"[bold red]{title}[/bold red]",
+                subtitle=subtitle,
+                title_align="left",
+                subtitle_align="left",
+                border_style="red"))
+
+    def unauthorized(self):
+        self.api_error("401", "Unauthorized", hide_subtitle=True)
+
+    def api_error(self, error_code: Optional[str], title: str, hide_subtitle: bool = False):
+        DEFAULT_ERROR_MESSAGE = "Unknown error. Please contact support."
+        ERROR_MESSAGE = ERROR_CODES.get(error_code, None) if error_code else None
+
+        if not ERROR_MESSAGE:
+            hide_subtitle = True
+
+        error_message = ERROR_MESSAGE if ERROR_MESSAGE else DEFAULT_ERROR_MESSAGE
+
+        self.print(
+            Panel(
+                f"[red]{title}[/red]\n\n"
+                f"[dim]Error message:[/dim]\n{error_message}",
+                title=f"[bold red]{PANEL_TITLE_ERROR} - {error_code}[/bold red]",
+                subtitle=f"[dim]Docs: https://docs.pipecat.cloud/troubleshooting/#{error_code}[/dim]" if not hide_subtitle else None,
+                title_align="left",
+                subtitle_align="left",
+                border_style="red"))
 
 
-def print_api_error(error_code: Optional[str], title: str):
-    error_message = "Unknown error" if not error_code else ERROR_CODES.get(
-        error_code, "Unknown error")
+def print_api_error(error_code: Optional[str], title: str, hide_subtitle: bool = False):
+    print("NOT IMPLEMENT REMOVE ME")
 
-    console.print(Panel(
-        f"[red]{title}[/red]\n\n"
-        f"[dim]Error message:[/dim]\n{error_message}",
-        title=f"[bold red]{PANEL_TITLE_ERROR} - {error_code}[/bold red]",
-        subtitle=f"[dim]Docs: https://docs.pipecat.cloud/troubleshooting/#{error_code}[/dim]",
-        title_align="left",
-        subtitle_align="left",
-        border_style="red"
-    ))
+
+console = PipecatConsole()