+++ src/pipecatcloud/cli/auth.py
@@ -7,14 +7,22 @@ from typing import AsyncGenerator, Optional, Tuple
 import aiohttp
 import typer
 from loguru import logger
-from rich.console import Console
+from rich.live import Live
 from rich.panel import Panel
 
 from pipecatcloud._utils.async_utils import synchronize_api, synchronizer
 from pipecatcloud._utils.auth_utils import requires_login
+from pipecatcloud._utils.console_utils import console
 from pipecatcloud._utils.http_utils import construct_api_url
-from pipecatcloud.cli import PANEL_TITLE_ERROR, PANEL_TITLE_SUCCESS
-from pipecatcloud.config import _remove_user_config, _store_user_config, config, user_config_path
+from pipecatcloud.api import API
+from pipecatcloud.config import (
+    _remove_user_config,
+    _store_user_config,
+    config,
+    dashboard_host,
+    user_config_path,
+)
+from pipecatcloud.exception import ConfigError
 
 auth_cli = typer.Typer(name="auth", help="Manage Pipecat Cloud credentials.", no_args_is_help=True)
 
@@ -80,127 +88,121 @@ def _open_url(url: str) -> bool:
         return False
 
 
-async def _set_credentials(
+def _set_credentials(
     token: str,
     account_org: str,
-):
-    console = Console()
-    with console.status("Storing user credentials", spinner="dots"):
+) -> bool:
+    try:
         _store_user_config(token, account_org)
+        return True
+    except ConfigError:
+        console.error(
+            f"Unable to store user config in path {user_config_path}. Is this folder write enabled?"
+        )
+    except Exception as e:
+        raise e
+    return False
 
 
 async def _get_account_org(
-        token: str, active_org: Optional[str] = None) -> Optional[Tuple[str, str]]:
-    console = Console()
-    # Retrieve account organization
-    with console.status("[dim]Obtaining account organization data[/dim]", spinner="dots"):
-        async with aiohttp.ClientSession() as session:
-            async with session.get(
-                f"{construct_api_url('organization_path')}",
-                headers={"Authorization": f"Bearer {token}"},
-            ) as resp:
-                if resp.status == 200:
-                    data = await resp.json()
-                    organizations = data["organizations"]
+        token: str, active_org: Optional[str] = None) -> Tuple[Optional[str], Optional[str]]:
 
-                    # If active_org is specified, try to find it in the list
-                    if active_org:
-                        for org in organizations:
-                            if org["name"] == active_org:
-                                return org["name"], org["verboseName"]
+    async with aiohttp.ClientSession() as session:
+        async with session.get(
+            f"{construct_api_url('organization_path')}",
+            headers={"Authorization": f"Bearer {token}"},
+        ) as resp:
+            if resp.status == 200:
+                data = await resp.json()
+                organizations = data["organizations"]
 
-                    # Default to first organization if active_org not found or not specified
-                    if organizations:
-                        return organizations[0]["name"], organizations[0]["verboseName"]
+                # If active_org is specified, try to find it in the list
+                if active_org:
+                    for org in organizations:
+                        if org["name"] == active_org:
+                            return org["name"], org["verboseName"]
 
-                    return None
-                else:
-                    raise Exception(f"Failed to retrieve account organization: {resp.status}")
+                # Default to first organization if active_org not found or not specified
+                if organizations:
+                    return organizations[0]["name"], organizations[0]["verboseName"]
 
-
-# ----- Login
-
-async def _login(active_org: Optional[str] = None):
-    console = Console()
-    auth_flow = _AuthFlow()
-
-    try:
-        async with auth_flow.start() as (token_flow_id, web_url):
-            if web_url is None:
-                console.print(
-                    Panel(
-                        "Unable to connect to Pipecat Cloud API. Please check your network connection and try again.",
-                        title=f"[red]{PANEL_TITLE_ERROR}[/red]",
-                        title_align="left",
-                        border_style="red",
-                    ))
-                return
-
-            with console.status("Waiting for authentication in the web browser", spinner="dots"):
-                # Open the web url in the browser
-                if _open_url(web_url):
-                    console.print(
-                        Panel(
-                            "The web browser should have opened for you to authenticate with Pipecat Cloud.\n"
-                            "If it didn't, please copy this URL into your web browser manually:\n\n"
-                            f"[blue][link={web_url}]{web_url}[/link][/blue]\n", ))
-                else:
-                    console.print("[dim]Unable to launch web browser[/dim]")
-                    console.print(
-                        "[bold]Please visit the following URL to confirm login[/bold]\n\n")
-                    console.print(web_url)
-
-            with console.status("Waiting for token flow to complete...", spinner="dots") as status:
-                for attempt in itertools.count():
-                    result = await auth_flow.finish()
-                    if result is not None:
-                        break
-                    status.update(f"Waiting for token flow to complete... (attempt {attempt + 1})")
-                if result is None:
-                    console.print(Panel(
-                        "[red]Authentication failed[/red]",
-                        title=f"[red]{PANEL_TITLE_ERROR}[/red]",
-                        title_align="left",
-                        border_style="red",
-                    ))
-                    return
-
-            # Retrieve user namespace
-            if not active_org:
-                try:
-                    account_name, account_name_verbose = await _get_account_org(result)
-                    logger.debug(f"Setting namespace to {account_name_verbose}")
-                    if account_name is None:
-                        raise
-                except Exception:
-                    console.print(
-                        Panel(
-                            "[red]Account has no associated namespace. Have you completed the onboarding process? Please first sign in via the web dashboard (https://dashboard.pipecat.cloud).[/red]",
-                            title=f"[red]{PANEL_TITLE_ERROR}[/red]",
-                            title_align="left",
-                            border_style="red",
-                        ))
-                    return
+                return None, None
             else:
-                account_name = active_org
+                raise Exception(f"Failed to retrieve account organization: {resp.status}")
 
-            console.print(Panel(
-                "[green]Web authentication finished successfully![/green]\n"
-                f"Account details stored to [magenta]{user_config_path}[/magenta]",
-                title=f"[green]{PANEL_TITLE_SUCCESS}[/green]",
-                title_align="left",
-                border_style="green",
-            ))
-            await _set_credentials(result, account_name)
-    except Exception:
-        pass
+# ---- Login ----
 
 
 @auth_cli.command(name="login", help="Login to Pipecat Cloud and get a new token")
 @synchronizer.create_blocking
 async def login(ctx: typer.Context):
     active_org = ctx.obj["org"]
-    await _login(active_org)
+    auth_flow = _AuthFlow()
+
+    try:
+        async with auth_flow.start() as (token_flow_id, web_url):
+            if web_url is None:
+                console.error(
+                    "Unable to connect to Pipecat Cloud API. Please check your network connection and try again.")
+                return
+
+            with Live(console.status("[dim]Waiting for authentication to complete...[/dim]", spinner="dots"), transient=True) as live:
+                # Open the web url in the browser
+                if _open_url(web_url):
+                    live.update(
+                        console.status(
+                            Panel(
+                                "The web browser should have opened for you to authenticate with Pipecat Cloud.\n"
+                                "If it didn't, please copy this URL into your web browser manually:\n\n"
+                                f"[blue][link={web_url}]{web_url}[/link][/blue]\n", )))
+                else:
+                    # For headless use-cases, just print the URL
+                    live.stop()
+                    console.print("[dim]Unable to launch web browser[/dim]")
+                    console.print(
+                        "[bold]Please visit the following URL to confirm login[/bold]\n")
+                    console.print(web_url)
+                live.start()
+                for attempt in itertools.count():
+                    result = await auth_flow.finish()
+                    if result is not None:
+                        break
+                    live.update(
+                        console.status(f"[dim]Waiting for authentication to complete... (attempt {attempt + 1})[/dim]"))
+                if result is None:
+                    live.stop()
+                    console.error("Authentication failed")
+                    return
+
+                live.update(
+                    console.status(
+                        "[dim]Obtaining account data[/dim]",
+                        spinner="dots"))
+
+                # Retrieve user namespace
+                if not active_org:
+                    try:
+                        account_name, account_name_verbose = await _get_account_org(result)
+                        live.stop()
+                        logger.debug(
+                            f"Setting namespace to {account_name_verbose} ({account_name})")
+                        if account_name is None:
+                            raise
+                    except Exception:
+                        live.stop()
+                        console.error(
+                            "Account has no associated namespace. Have you completed the onboarding process? Please first sign in via the web dashboard.")
+                        return
+                else:
+                    account_name = active_org
+
+            if _set_credentials(result, account_name):
+                console.success(
+                    "Web authentication finished successfully!\n"
+                    f"[dim]Account details stored to [magenta]{user_config_path}[/magenta][/dim]"
+                )
+    except Exception:
+        pass
 
 
 # ----- Logut
@@ -209,34 +211,41 @@ async def login(ctx: typer.Context):
 @synchronizer.create_blocking
 @requires_login
 async def logout():
-    """
-    # This command should delete the session token
-    # due to networkless token verification, the token will not be invalidated
-    # it will just be removed from the local config file
-    # if we want token invalidation, we'll need to either use sign-in tokens with clerk
-    # or use a networked token verification system of our own design
-    """
-    console = Console()
-
     with console.status("Removing user ID", spinner="dots"):
         _remove_user_config()
 
-    console.print(Panel(
-        "You are now logged out of Pipecat Cloud.",
-        title=f"[yellow]{PANEL_TITLE_SUCCESS}[/yellow]",
-        title_align="left",
-        border_style="yellow",
-    ))
+    console.success(
+        "User credentials for Pipecat Cloud removed. Please sign out via dashboard to fully revoke session.",
+        subtitle=f"[dim]Please visit:[/dim] {dashboard_host}/sign-out")
 
 
 @auth_cli.command(name="whoami", help="Display data about the current user.")
 @synchronizer.create_blocking
 @requires_login
 async def whomai():
-    console = Console()
-    token = config.get("token")
-    active_org = config.get("org")
+    with Live(console.status("[dim]Requesting current user data...[/dim]", spinner="dots"), transient=True) as live:
+        try:
+            user_data = await API.whoami()
 
+            live.update(
+                console.status("[dim]Requesting user namespace / organization data...[/dim]"))
+
+            # Retrieve default user organization
+            account_name, account_name_verbose = await API.organizations_current()
+
+            if not account_name or not account_name_verbose:
+                raise
+
+            live.stop()
+
+            console.success(
+                f"[bold]User ID:[/bold] {user_data['user']['userId']}\n"
+                f"[bold]Active Organization:[/bold] {account_name_verbose} [dim]({account_name})[/dim]", )
+        except Exception:
+            live.stop()
+            console.error("Unable to retrieve user data. Please contact support.")
+
+    """
     async with aiohttp.ClientSession() as session:
         try:
             # Retrieve user data from whoami endpoint
@@ -271,3 +280,4 @@ async def whomai():
                 border_style="red",
             ))
             return
+    """